isolate the logic of settings/general from the rendering
